Fixed segfault when input file has not been found
file_in is initialized with stdin. When opening it, a failure is
indicated by the NULL value assigned to file_in. The application
then tries to close the NULL file descriptor as it is not equal to
stdin which gives a segmentation fault.

This patch adds a check if the file descriptor is NULL and does not
try to close it if this is the case.